+++ src/host/coreaudio/macos/enumerate.rs
@@ -81,7 +81,6 @@ impl Iterator for Devices {
     fn next(&mut self) -> Option<Device> {
         self.0.next().map(|id| Device {
             audio_device_id: id,
-            is_default: false,
         })
     }
 }
@@ -109,10 +108,7 @@ pub fn default_input_device() -> Option<Device> {
         return None;
     }
 
-    let device = Device {
-        audio_device_id,
-        is_default: true,
-    };
+    let device = Device { audio_device_id };
     Some(device)
 }
 
@@ -139,10 +135,7 @@ pub fn default_output_device() -> Option<Device> {
         return None;
     }
 
-    let device = Device {
-        audio_device_id,
-        is_default: true,
-    };
+    let device = Device { audio_device_id };
     Some(device)
 }