+++ src/host/coreaudio/macos/mod.rs
@@ -159,18 +159,22 @@ impl DeviceTrait for Device {
 #[derive(Clone, PartialEq, Eq)]
 pub struct Device {
     pub(crate) audio_device_id: AudioDeviceID,
-    is_default: bool,
+}
+
+fn is_default_device(device: &Device) -> bool {
+    default_input_device()
+        .map(|d| d.audio_device_id == device.audio_device_id)
+        .unwrap_or(false)
+        || default_output_device()
+            .map(|d| d.audio_device_id == device.audio_device_id)
+            .unwrap_or(false)
 }
 
 impl Device {
     /// Construct a new device given its ID.
     /// Useful for constructing hidden devices.
     pub fn new(audio_device_id: AudioDeviceID) -> Self {
-        Device {
-            audio_device_id,
-            // TODO: This could be made to detect the default device properly.
-            is_default: false,
-        }
+        Self { audio_device_id }
     }
 
     fn name(&self) -> Result<String, DeviceNameError> {
@@ -523,7 +527,7 @@ where
 }
 
 fn audio_unit_from_device(device: &Device, input: bool) -> Result<AudioUnit, coreaudio::Error> {
-    let output_type = if device.is_default && !input {
+    let output_type = if is_default_device(device) && !input {
         coreaudio::audio_unit::IOType::DefaultOutput
     } else {
         coreaudio::audio_unit::IOType::HalOutput
@@ -665,7 +669,7 @@ impl Device {
 
         // If we didn't request the default device, stop the stream if the
         // device disconnects.
-        if !self.is_default {
+        if !is_default_device(self) {
             add_disconnect_listener(&stream, error_callback_disconnect)?;
         }
 
@@ -770,7 +774,7 @@ impl Device {
 
         // If we didn't request the default device, stop the stream if the
         // device disconnects.
-        if !self.is_default {
+        if !is_default_device(self) {
             add_disconnect_listener(&stream, error_callback_disconnect)?;
         }